+++ lib/readymade/instant_form.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 module Readymade
-  VERSION = '0.1.4'
+  class InstantForm < Form
+
+  end
 end