+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 Changelog
 All notable changes to this project will be documented in this file.
 
-[0.1.3] - 2021-11.20
+## [0.1.4] - 2021-11-22
 
-Improvements
+### Features
+
+* Add `Readymade::InstantForm`
+
+### Improvements
+
+* Call `build_form` inside `form_valid?` if `@form` is not defined
+
+## [0.1.3] - 2021-11-20
+
+### Improvements
 
 * Fix error when required param for instant form is not defined